requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
python-dotenv==1.0.0
markdown==3.5
python-wordpress-xmlrpc==2.3
selenium==4.12.0
